requirements.txt
scikit-learn >= 0.23.2
requests >= 2.25.0
beautifulsoup4 >= 4.9.3
asyncio >= 3.4.3
aiohttp >= 3.7.3
pymongo >= 3.11.1
dnspython >= 2.0.0
motor >= 2.3.0
lxml >= 4.6.2
scikit-surprise >= 1.1.1
numpy >= 1.19.4
pandas >= 1.1.4